arm: Split PagedResponse to a separate file
